rril: address comments from patch 174947
BZ: 196237

This patch addresses comments made during review
on patch 174947

+++ CORE/ND/te.h
@@ -781,7 +781,7 @@ public:
 
     CEvent* GetRadioStateChangedEvent() { return m_pRadioStateChangedEvent; }
 
-    UINT32 GetCnapCniValidity() { return m_uiCnapCniValidity; }
+    int GetCnapCniValidity() { return static_cast<int>(m_uiCnapCniValidity); }
     void SetCnapCniValidity(UINT32 uiValidity) { m_uiCnapCniValidity = uiValidity; }
 
     char* GetCnapName() { return m_szCnapName; }
